Fixed errors introduced by upgrading the Eigen dependency to version 3.4.0

See merge request NumCSE/NumCSE!103

+++ LectureCodes/Filtering/blur/Eigen/blur.hpp
@@ -4,24 +4,24 @@ using Eigen::MatrixXd;
 
 /* SAM_LISTING_BEGIN_0 */
 MatrixXd blur(const MatrixXd &P, const MatrixXd &S) {
-  const long m = P.rows(), n = P.cols(), M = S.rows(), N = S.cols(),
+  const int m = P.rows(), n = P.cols(), M = S.rows(), N = S.cols(),
              L = (M - 1) / 2;
   if (M != N) {
     std::cout << "Error: S not quadratic!\n";
   }
 
   MatrixXd C(m, n);
-  for (long l = 1; l <= m; ++l) {
-    for (long j = 1; j <= n; ++j) {
+  for (int l = 1; l <= m; ++l) {
+    for (int j = 1; j <= n; ++j) {
       double s = 0;
-      for (long k = 1; k <= (2 * L + 1); ++k) {
-        for (long q = 1; q <= (2 * L + 1); ++q) {
-          double kl = l + k - L - 1;
+      for (int k = 1; k <= (2 * L + 1); ++k) {
+        for (int q = 1; q <= (2 * L + 1); ++q) {
+          int kl = l + k - L - 1;
           if (kl < 1)
             kl += m;
           else if (kl > m)
             kl -= m;
-          double jm = j + q - L - 1;
+          int jm = j + q - L - 1;
           if (jm < 1)
             jm += n;
           else if (jm > n)
